Finish the JavaScript variables and operators lesson.
Edit the .js file to declare and operate on various variables while using "Live Preview" extension on VSCode to view code.

Note: Console in Live Preview is finicky and doesn't seem to remember any declared variables. Should check if this is the case with devtools on Google Chrome as well, or just an issue with Live Preview.